Cleaned up

+++ Page.py
@@ -38,10 +38,10 @@ class Page (Sprite):
     item_height = self.height / self.col_length
 
     # determine the dimensions of the interior game element
-    game_element_x_pos = item_width / 8
-    game_element_y_pos = item_height / 8
-    game_element_width = item_width * 0.75
-    game_element_height = item_height * 0.75
+    game_element_x_pos = int(item_width / 8)
+    game_element_y_pos = int(item_height / 8)
+    game_element_width = int(item_width * 0.75)
+    game_element_height = int(item_height * 0.75)
 
     # used to calculate new elements position during population
     x_pos = 0
